Load-balancer configuration code
This commit adds a daemon that listens for events from an OpenShift
Enterprise installation that is configured to publish gear creation and
deletion events ot ActiveMQ.  The daemon configures a load balancer to
route connections directly through the port-proxy, bypassing gear-level
haproxy instances.

This commit also adds a command-line utility for manipulating the load
balancer as well as common code that is used by both the command-line
utility and the daemon code.