LBaaS model: require host in authenticate method
Because the keystone host is likely to be different from the LBaaS host,
require an explicit host to be provided to the authenticate method
rather than defaulting to the post that was provided to the initializer.

Require a host argument in initialize.  In the F5 model, the host
provided to authenticate is assigned to @host, so it makes sense for the
host to be optional for the initializer; the same is not true for LBaaS.

@user and @passwd are not used outside authenticate, so continue to
allow them to be omitted in initialize.  (If they are provided to
initialize, the provided arguments are used as the default values in
authenticate, so they are optional for authenticate if they are provided
to initialize.)