controllers/lbaas.rb: Don't dequeue submitted ops
Operations should remain in @ops even after they have been submitted to
the load balancer.  We know that an operation has been submitted when it
has a non-empty @jobids attribute, and we check @ops for operations that
have been submitted so that we can poll LBaaS do find out when
operations are completed.  If we immediately remove operations from @ops
when we submit them, we never poll them, and they continue to block
other operations that depend on them.