Move control of monitor_name to daemon
Drop BIGIP_MONITOR and LBAAS_MONITOR from the default load-balancer.conf
and add MONITOR_NAME.

Modify the controllers to take a monitor_name argument for their
create_pool methods.

Instead of having the controllers read backend-specific configuration
settings for the monitor name, have the daemon read a setting for the
monitor name and pass the value down to the controllers.

Allow a monitor name to be specified using oo-admin-load-balancer