Allow per-app (or per-domain) monitors
The value of MONITOR_NAME in load-balancer.conf may now include format
codes %a and %n, for which the daemon will substitute the application's
name and namespace, respectively, when creating a monitor.

Add a MONITOR_PATH configuration setting to load-balancer.conf which may
be used to specify the health-check path to be used when creating
a monitor.  The same format codes can be used in MONITOR_PATH that can
be used in MONITOR_NAME.

Add a @monitors attribute to the controllers.

In all controllers, get the list of monitors from the load balancer and
store them into the new @monitors attribute in the initialize methods.

Add create_monitor and delete_monitor methods to all controllers and
models, and add get_monitor_names methods to all models.

The daemon will now create a new monitor if one with the given name does
not already exist.  One can thus take advantage of embedded formats in
MONITOR_NAME to create a new monitor for each application or namespace.