Move "/Common/" partition prefix to F5 code
Instead of including "/Common/" in the pool and route names the daemon
gives to the F5 controller, have the F5 model append the "/Common/"
prefix.

The "/Common/" prefix identifies the partition in which the pool or
route is created, and because the concept of a partition is F5-specific,
the prefix logically belongs in the F5-specific code.  LBaaS does not
like the prefix, and so we must not include it in the general code.