Rewrite LBaaS keystone authentication
The LBaaS tenant and the Keystone tenant are not necessarily the same,
so add a new configuration option LBAAS_KEYSTONE_TENANT in addition to
the old LBAAS_TENANT option, and use its value for keystone authentication.

The username and password are only used for keystone authentication, so
rename the configuration options LBAAS_USERNAME to LBAAS_KEYSTONE_USERNAME
and LBAAS_PASSWORD to LBASS_KEYSTONE_PASSWORD.

Update the README and default .conf file accordingly.