models/lbaas.rb: Refactor REST Calls
Add new methods get, post, and put to LBaaSLoadBalancerModel that wrap
RestClient::Request.execute, and use these methods instead of directly
using RestClient.get, RestClient.post, and RestClient.put.

Introduce a new method lbaas_headers that returns the headers for LBaaS
REST calls, and use this method in the new get, post, and put methods to
reduce code repetition.

The new lbaas_headers method sends an X-Auth-Token: header with whatever
value is in @keystone_token if that value is not nil, so set
@keystone_token to nil at the start of the authenticate method before
the first request to get the temporary token, and store this temporary
token in @keystone_token before making the request that uses it to get
the permanent token.