LBaaS: re-authenticate when the token expires
During authentication with keystone, store the information needed for
authentication as well as the expiration time of the permanent token.

Add a method maybe_authenticate to the model that checks this expiration
time and calls the authenticate method using the stored authentication
information if the token has expired.

In the controller, call the model's maybe_authenticate method at the
start of the update method.