controllers/lbaas.rb: Don't reap a cancelled op
When submitting an operation to LBaaS in the update method, call
reap_op_if_no_remaining_tasks in the begin part of the begin/rescue/end
block rather than after the begin/rescue/end block so as to avoid reaping
a job that has already been cancelled.

When an Operation was reaped after it had already been cancelled, it could
cause the blocked_on_cnt of operations it had been blocking to be
decremented an extra time, which would cause blocked_on_cnt to go negative.
Because an Operation is considered ready when its jobids is empty and its
blocked_on_cnt is zero, this bug could cause Operation objects never to be
considered ready.