Make model initialize methods more consistent
Add an initializer for the F5 model that takes host, username, and
password, and make the F5 controller pass @bigip_host, @bigip_username,
and @bigup_password to the F5 model's initializer for consistency with
the dummy model and to simplify future changes.

Make the dummy model's initialize method's arguments non-optional.