daemon.rb: delete_pool even if delete_route fails
When an application is deleted, the daemon first tries to delete the
corresponding routing rule and then tries to delete the corresponding
pool.  This commit wraps these two steps in a begin/ensure/end block so
that the daemon tries to delete the pool even if there is an exception
from delete_route.