Make monitor up code configurable
daemon.rb: Read the new MONITOR_UP_CODE setting from load-balancer.conf,
and when creating new monitors, use the value of that setting as the
code the monitor should take to indicate that a pool member is up.

Update the README.md file and shipped load-balancer.conf file with this
new setting.