Make monitor type configurable
daemon.rb: Read the MONITOR_TYPE setting, which should be one either
'http-ecv' or 'https-ecv', from load-balancer.conf and use the value of
this setting when creating new monitors.

controllers/*, models/*: Add a "type" parameter to create_monitor methods,
and pass the specified type to the model when creating a monitor.

Update the README.md file and shipped load-balancer.conf file with this
new setting.