Make monitor interval configurable
daemon.rb: Read the MONITOR_INTERVAL setting from load-balancer.conf and
use the value of this setting when creating new monitors.

controllers/*, models/*: Add an "interval" parameter to create_monitor
methods, and pass the specified interval to the model when creating
monitors.

Update the README.md file and shipped load-balancer.conf file with this
new setting.