Make monitor timeout configurable
daemon.rb: Read the MONITOR_TIMEOUT setting from load-balancer.conf and
use the value of this setting when creating new monitors.

controllers/*, models/*: Add an "timeout" parameter to create_monitor
methods, and pass the specified timeout to the model when creating
monitors.

Update the README.md file and shipped load-balancer.conf file with this
new setting.