Delete monitor on app delete if it's not shared
daemon.rb: When deleting an application, if the application has a monitor
and the monitor is specific to that application, then delete that monitor.

To determine whether the monitor is specific to the application, use the
heuristic of checking whether the template for the name includes the
formats for application name and namespace.