controllers/lbaas.rb: Fix delete_monitor
Fix the scheduling of delete_monitor operations with LBaaS and the
asynchronous controller.

controllers/f5.rb, controllers/lbaas.rb, controllers/load_balancer.rb: Add
an optional pool_name parameter to the delete_monitor method.

daemon.rb: In the delete_application method, when deleting the monitor,
pass in the name of the associated pool.

controllers/lbaas.rb: In the delete_monitor method, if a pool_name is
specified and there is a pending delete_pool operation for that pool, block
the delete_monitor operation on the delete_pool operation.